backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Group 1", email = "example@example.com"},
]
dependencies = [
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "supabase>=2.9.1",
    "fastapi>=0.115.4",
    "uvicorn>=0.32.0",
    "authlib>=1.3.2",
    "starlette>=0.41.3",
    "itsdangerous>=2.2.0",
    "python-jose>=3.3.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "sqlmodel>=0.0.22",
    "aiosqlite>=0.20.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = "uvicorn src.main:app --reload"
